Move PreviewTerminateSector(s) code out of SDK interface and update lotus dep
